Update snippet list view to show multiple fields
The snippet list view can now show multiple fields of a snippet (instead of using the string representation); the fields to show are defined by the list_fields property of the snippet class (an iterable of strings). The first field will be the selection link (other fields are currently just for informational purposes - sorting is not supported currently).