Embedded snippet support
Added support for embedding snippets in rich-text content.

Snippets are rendering using the <app_label>/<snippet_model>.html file (with a fallback to wagtailcore/snippet.html).

Each snippet type has a button added to the rich-text editor to enable selection of snippets.

The DB stored content for snippets is of the form:

<embed embedtype="snippet" id="42" snippet-type="demo.snippet_type" />